package/requirements.txt
pillow==10.2.0
boto3==1.34.0
